Move utils and common into a core package
