+++ aces/idt/__init__.py
@@ -1,4 +1,4 @@
-from .common import (
+from .core.common import (
     clf_processing_elements,
     error_delta_E,
     generate_reference_colour_checker,
@@ -6,8 +6,8 @@ from .common import (
     optimisation_factory_Oklab,
     png_compare_colour_checkers,
 )
-from .prosumer_camera import IDTGeneratorProsumerCamera
-from .utilities import (
+from .generators.prosumer_camera import IDTGeneratorProsumerCamera
+from .core.utilities import (
     hash_file,
     list_sub_directories,
     mask_outliers,
